requirements.txt
anthropic>=0.96.0
psutil>=5.9.0
ccxt>=4.5.49
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.20.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.8.0
scipy>=1.11.0
streamlit>=1.35.0
ta>=0.11.0
